% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtgazer.R
\name{gtgazer}
\alias{gtgazer}
\title{Create a stargazer-style regression table using gt}
\usage{
gtgazer(
  ...,
  dependent_variables = NULL,
  independent_variables = everything(),
  summary_statistics = gtgazer_default_summary_statistics(),
  additional_coefficients = NULL,
  additional_summaries = NULL,
  model_names = NULL,
  model_types = NULL,
  include_model_types = NULL,
  include_model_numbers = TRUE,
  decimal_digits_max = 3,
  decimal_mark = ".",
  keep_models = TRUE
)
}
\arguments{
\item{...}{One or more model objects}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Create a regression table in the style of the
\link[=https://cran.r-project.org/web/packages/stargazer/index.html]{stargazer}
package, powered by \link[gt:gt-package]{gt}.
}
\examples{
example_lm <- lm(mpg ~ wt + hp, data = mtcars)
example_glm <- glm(vs ~ hp + cyl, data = mtcars, family = binomial)
example_clm <- ordinal::clm(as.ordered(cyl) ~ hp + vs, data = mtcars)

gtgazer(example_lm, example_glm, example_clm)

gtgazer(
	example_lm, example_glm, example_clm,
	independent_variables = c(
		"Weight" = "wt",
		where(is_intercept),
		everything()
	),
	additional_coefficients = list(
		"Education fixed effects" = cli::symbol$tick,
		"Party ID fixed effects" = cli::symbol$tick,
		"Survey week fixed effects" = cli::symbol$tick
	)
)
}
