% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercept_column.R
\name{is_intercept}
\alias{is_intercept}
\title{Selection helper for model intercepts}
\usage{
is_intercept(x)
}
\arguments{
\item{x}{An independent variable in a \code{\link[=gtgazer]{gtgazer()}} table.}
}
\value{
A \code{\link{TRUE}}/\code{\link{FALSE}} value.
}
\description{
Selection helper for model intercepts
}
\examples{
# Move intercepts to the top of a table
gtgazer(
	lm(mpg ~ wt + hp, data = mtcars),
	ordinal::clm(as.ordered(cyl) ~ hp + vs, data = mtcars),
	independent_variables = c(
		where(is_intercept),
		everything()
	)
)
}
